requirements.txt
crewai
duckduckgo-search
python-dotenv
requests
beautifulsoup4
langchain_openai
flask
flask-caching
markdown
langchain-google-genai